move integration tests to test module
