+++ build.gradle.kts
@@ -1,4 +1,6 @@
 plugins {
+    alias(libs.plugins.androidApplication) apply false
     alias(libs.plugins.androidLibrary) apply false
+    alias(libs.plugins.androidTest) apply false
     alias(libs.plugins.kotlinAndroid) apply false
 }
