+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,7 +1,5 @@
 import org.gradle.api.JavaVersion
 
 object Versions {
-  const val androidLibraryVersion = "8.2.1"
-  const val kotlinVersion = "1.8.22"
   val javaVersion = JavaVersion.VERSION_17
 }
