+++ instrumentedtest/src/main/res/raw/change_image.png
