+++ instrumentedtest/src/main/res/raw/change_text.png
