+++ instrumentedtest/src/main/res/raw/dialog.png
