+++ instrumentedtest/src/main/res/raw/no_logo.png
