+++ instrumentedtest/src/main/res/raw/recycler_view.png
