+++ screencaptor-sample/application/build.gradle.kts
@@ -1,0 +1,22 @@
+plugins {
+  alias(libs.plugins.androidApplication)
+  alias(libs.plugins.kotlinAndroid)
+}
+
+apply(from = "../../gradle/android-module.gradle")
+
+android {
+  defaultConfig {
+    testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+  }
+
+  namespace = "com.wealthfront.screencaptor.sample"
+}
+
+dependencies {
+  api(project(":screencaptor-sample:impl"))
+  implementation(libs.recyclerview)
+  implementation(libs.espresso.core)
+  implementation(libs.espresso.contrib)
+  implementation(libs.appcompat)
+}
