+++ screencaptor-sample/impl/src/main/res/drawable/add_accounts.png
