+++ screencaptor-sample/impl/src/main/res/layout/screenshot_test.xml
@@ -1,0 +1,72 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<androidx.core.widget.NestedScrollView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/scrollView"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="#4B0082"
+    android:scrollbars="vertical"
+    tools:ignore="HardcodedText"
+    >
+
+  <LinearLayout
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:orientation="vertical"
+      >
+
+    <ImageView
+        android:id="@+id/wealthfrontIcon"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:padding="40dp"
+        android:src="@drawable/wf_logo"
+        />
+
+    <EditText
+        android:id="@+id/textField"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:gravity="center"
+        android:hint="Welcome to Wealthfront!!!"
+        android:inputType="text"
+        android:textColorHint="#FFFFFF"
+        tools:ignore="Autofill,LabelFor"
+        />
+
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Some sample data which is really long, so long that it wraps to another line and maybe even three lines"
+        />
+
+    <Button
+        android:id="@+id/showDialog"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Show dialog"
+        />
+
+    <ScrollView
+        android:id="@+id/anotherScrollView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        />
+
+    <HorizontalScrollView
+        android:id="@+id/horizonalScrollView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/messageList"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        />
+  </LinearLayout>
+
+</androidx.core.widget.NestedScrollView>
