+++ screencaptor-sample/impl/src/main/res/raw/dialog.png
