+++ screencaptor-sample/impl/src/test/java/com/wealthfront/screencaptor/sample/impl/ScreenshotTest.kt
@@ -1,0 +1,57 @@
+package com.wealthfront.screencaptor.sample.impl
+
+import android.graphics.BitmapFactory.decodeByteArray
+import android.graphics.BitmapFactory.decodeFile
+import androidx.test.core.app.launchActivity
+import androidx.test.espresso.Espresso
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.action.ViewActions.click
+import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.ext.junit.rules.ActivityScenarioRule
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry.getInstrumentation
+import com.wealthfront.screencaptor.ScreenCaptor
+import org.junit.After
+import org.junit.Assert.assertTrue
+import org.junit.Rule
+import org.junit.Test
+import org.junit.rules.TemporaryFolder
+import org.junit.runner.RunWith
+import org.robolectric.annotation.GraphicsMode
+
+@RunWith(AndroidJUnit4::class)
+@GraphicsMode(org.robolectric.annotation.GraphicsMode.Mode.NATIVE)
+class ScreenshotTest {
+
+  @Rule
+  @JvmField
+  val folder = TemporaryFolder()
+
+  @get:Rule
+  var activityTestRule: ActivityScenarioRule<SampleActivity> = ActivityScenarioRule(com.wealthfront.screencaptor.sample.impl.SampleActivity::class.java)
+
+  @After
+  fun cleanUpScreenshots() {
+    folder.root.deleteRecursively()
+  }
+
+  @Test
+  fun takeScreenshot_dialog() {
+    onView(withId(R.id.showDialog)).perform(click())
+
+    val activityScenario = launchActivity<SampleActivity>()
+
+    Espresso.onIdle()
+    ScreenCaptor.takeScreenshot(
+      activityScenario = activityScenario,
+      screenshotName = "screenshot_dialog",
+      screenshotDirectory = folder.root.path
+    )
+
+    val screenshot = folder.root.listFiles()!!.find { it.name.contains("screenshot_dialog") }!!
+    val actual = decodeFile(screenshot.path)
+    val expectedBytes = getInstrumentation().context.resources.openRawResource(R.raw.dialog).readBytes()
+    val expected = decodeByteArray(expectedBytes, 0, expectedBytes.size)
+    assertTrue(actual.sameAs(expected))
+  }
+}
