--- screencaptor-test/src/main/java/com/wealthfront/screencaptor/MessageAdapter.kt
@@ -1,28 +1,0 @@
-package com.wealthfront.screencaptor
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.wealthfront.screencaptor.test.R
-
-class MessageAdapter(private val messages: Array<String>) : RecyclerView.Adapter<MessageAdapter.MessageViewHolder>() {
-
-  override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MessageViewHolder {
-    val view = LayoutInflater.from(parent.context)
-      .inflate(R.layout.message_row_item, parent, false) as TextView
-
-    return MessageViewHolder(view)
-  }
-
-  override fun onBindViewHolder(holder: MessageViewHolder, position: Int) {
-    holder.view.text = messages[position]
-  }
-
-  override fun getItemCount(): Int = messages.size
-
-  data class MessageViewHolder(
-    val view: TextView
-  ) : RecyclerView.ViewHolder(view)
-}
-
