+++ screencaptor/src/test/res/drawable/wf_logo.png
