+++ screencaptor/src/test/res/layout/screenshot_test.xml
